+++ server/service/Services.php
@@ -106,7 +106,7 @@ class Services
 
         $mail = new Mailer($this->db, $this->transaction, $this->user_input, $this->router, $this->condition);        
 
-        $this->job_scheduler = new JobScheduler($this->db, $this->transaction, $mail, $this->condition);
+        $this->job_scheduler = new JobScheduler($this->db, $this->transaction, $mail, $this->condition, $this->user_input);
         $this->user_input->setJobSchedulerService($this->job_scheduler);
         
         if ($fullMode) {